refactor (structure): change structure of code similar to generator output
