+++ template.json
@@ -5,6 +5,7 @@
     "react-redux-firebase": "3.0.6",
     "firebase": "^7.1.0",
     "redux-firestore": "latest",
+    "react-router-dom": "^5.1.2",
     "@testing-library/react": "^9.3.2",
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/user-event": "^7.1.2"
