+++ template/src/containers/App/Router.js
@@ -1,0 +1,17 @@
+import React from 'react';
+import { BrowserRouter} from "react-router-dom";
+import createRoutes from '../../routes/index';
+
+const Routes = createRoutes();
+
+function Router() {
+  return (
+    <div>
+      <BrowserRouter>
+        {Routes}
+      </BrowserRouter>
+    </div>
+  )
+}
+
+export default Router;
