+++ template/src/routes/Home/components/Home/tests/Home.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render } from '@testing-library/react';
-import Home from './Home';
+import Home from '../Home';
 
 test('renders learn react link', () => {
   const { getByText } = render(<Home />);
