+++ template/src/routes/Home/index.js
@@ -1,0 +1,7 @@
+import Home from './components/Home'
+
+// Sync route definition
+export default {
+  path: '/',
+  component: Home
+}
