+++ template/src/store/reducers.js
@@ -2,7 +2,6 @@ import { combineReducers } from 'redux'
 import { firestoreReducer as firestore } from 'redux-firestore'
 import { firebaseReducer as firebase} from 'react-redux-firebase'
 
-
 const rootReducer = combineReducers({
   firestore,
   firebase,
